experimenting with spring boot config
