+++ spring-cloud-gcp-samples/spring-cloud-gcp-data-spanner-sample/src/main/java/com/example/WebController.java
@@ -29,7 +29,6 @@ import org.springframework.web.bind.annotation.RestController;
  * @author Chengyuan Zhao
  */
 @RestController
-@EnableSpannerRepositories(namedQueriesLocation = "classpath:/spanner-named-queries.properties")
 public class WebController {
 
 	@Autowired
