+++ spring-cloud-gcp-samples/spring-cloud-gcp-data-spanner-sample/src/main/resources/application.properties
@@ -1,3 +1,2 @@
-spring.cloud.gcp.spanner.projectId=<your-project-id>
 spring.cloud.gcp.spanner.instanceId=spring-demo
 spring.cloud.gcp.spanner.database=trades
